np-41723: moved interfaces to commons
