+++ cristin-commons/src/main/java/no/unit/nva/cristin/common/client/IClientProvider.java
@@ -1,4 +1,4 @@
-package no.unit.nva.cristin.organization;
+package no.unit.nva.cristin.common.client;
 
 public interface IClientProvider<T> {
