+++ cristin-commons/src/main/java/no/unit/nva/cristin/common/client/IQueryApiClient.java
@@ -1,4 +1,4 @@
-package no.unit.nva.cristin.organization;
+package no.unit.nva.cristin.common.client;
 
 import no.unit.nva.cristin.model.SearchResponse;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
